+++ uikit/src/main/java/com/sendbird/uikit/fragments/PromoteOperatorsFragment.java
@@ -7,6 +7,7 @@ import android.view.View;
 import androidx.annotation.DrawableRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
 import androidx.annotation.StyleRes;
 import androidx.fragment.app.Fragment;
 
@@ -309,6 +310,32 @@ public class PromoteOperatorsFragment extends SelectUserFragment {
          */
         public Builder setHeaderLeftButtonListener(View.OnClickListener listener) {
             this.headerLeftButtonListener = listener;
+            return this;
+        }
+
+        /**
+         * Sets the icon when the data is not exists.
+         *
+         * @param resId the resource identifier of the drawable.
+         * @param tint Color state list to use for tinting this resource, or null to clear the tint.
+         * @return This Builder object to allow for chaining of calls to set methods.
+         * @since 2.1.6
+         */
+        public Builder setEmptyIcon(@DrawableRes int resId, @Nullable ColorStateList tint) {
+            bundle.putInt(StringSet.KEY_EMPTY_ICON_RES_ID, resId);
+            bundle.putParcelable(StringSet.KEY_EMPTY_ICON_TINT, tint);
+            return this;
+        }
+
+        /**
+         * Sets the text when the data is not exists
+         *
+         * @param resId the resource identifier of text to be displayed.
+         * @return This Builder object to allow for chaining of calls to set methods.
+         * @since 2.1.6
+         */
+        public Builder setEmptyText(@StringRes int resId) {
+            bundle.putInt(StringSet.KEY_EMPTY_TEXT_RES_ID, resId);
             return this;
         }